clk: Add fixed-clock-nvmem provider
Adds a fixed-rate clock that retrieves its rate from an NVMEM provider.
This allows to store clock settings in EEPROM or EFUSE or similar device.

Component shortages lead to boards being shipped with different clock
crystals, based on what was available at the time. The clock frequency
was written to EEPROM at production time. Systems can adapt to a wide
range of input frequencies using the clock framework, but this required
us to patch the devicetree at runtime or use some custom driver. This
provides a more generic solution.

Because this clock depends on other hardware (typical NVMEM provider is
an I2C EEPROM) it cannot be integrated into clk-fixed which uses
CLK_OF_DECLARE to initialize.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>

Series-to: linux-clk@vger.kernel.org
Series-version: 4

Series-changes: 3
Change compatible to fixed-clock-nvmem

[rgantois: rebased from v6.1.112 to v6.6.58]

%% original patch: 0020-clk-Add-fixed-clock-nvmem-provider.patch