netlink: specs: Expand the pse netlink command with PoE interface
Update the pse attributes prefix to be able to add PoE interface.
To keep the standard naming in the netlink spec I remove the "podl" part of
the attribute prefix naming. This way we can select either PoDL attribute
which have podl in the attribute name or the PoE attribute.

Example usage:
./ynl/cli.py --spec netlink/specs/ethtool.yaml --no-schema --do pse-get \
             --json '{"header":{"dev-name":"eth0"}}'
{'header': {'dev-index': 4, 'dev-name': 'eth0'},
 'pse-admin-state': 3,
 'pse-pw-d-status': 4}

./ynl/cli.py --spec netlink/specs/ethtool.yaml --no-schema --do pse-set \
             --json '{"header":{"dev-name":"eth0"}, "pse-admin-control":3}'

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>