net: netdevsim: ptp_mock: Convert to netdev_ptp_clock_register
The hardware registration clock for net device is now using
netdev_ptp_clock_register to save the net_device pointer within the PTP
clock xarray. netdevsim is registering its ptp through the mock driver.
It is the only driver using the mock driver to register a ptp clock.
Convert the mock driver to the new API.

Reviewed-by: Rahul Rameshbabu <rrameshbabu@nvidia.com>
Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
---
Shouldn't we move the mock driver in the netdevsim directory as it is only
used by netdevsim driver?

Changes in v8:
- New patch