net: pse-pd: Add support for reporting events
Add support for devm_pse_irq_helper() to register PSE interrupts. This aims
to report events such as over-current or over-temperature conditions
similarly to how the regulator API handles them but using a specific PSE
ethtool netlink socket.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
---

Dropped the Reviewed-by on this patch as there are a lot of modification.

Change in v2:
- Add support for PSE ethtool notification.
- Saved the attached phy_device in the pse_control structure to know who's
  interface should have the notification.
- Rethink devm_pse_irq_helper() without devm_regulator_irq_helper() call.