net: ethtool: tsconfig: Fix ts filters and types enums size check
Fix the size check for the hwtstamp_tx_types and hwtstamp_rx_filters
enumerations. Align this check with the approach used in tsinfo for
consistency and correctness.

__HWTSTAMP_TX_CNT is matching the last HWTSTAMP_TX enum value + 1.
If __HWTSTAMP_TX_CNT is 33, HWTSTAMP_TX last possible value is 32.
BIT(32) will overflow u32 size and an error should be reported.
If __HWTSTAMP_TX_CNT is 32 HWTSTAMP_TX last possible value is 31.
BIT(31) is last bit of u32 type size therefore it is ok to work with
that.
This means that we should report an error only if __HWTSTAMP_TX_CNT if
greater than 32.

__HWTSTAMP_FILTER_CNT is similar.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
Fixes: 6e9e2eed4f39 ("net: ethtool: Add support for tsconfig command to get/set hwtstamp config")
---

Changes in v2:
- Update commit message.