net: Add support for providing the PTP hardware source in tsinfo
Multi-PTP source support within a network topology has been merged,
but the hardware timestamp source is not yet exposed to users.
Currently, users only see the PTP index, which does not indicate
whether the timestamp comes from a PHY or a MAC.

Add support for reporting the hwtstamp source using a
hwtstamp-source field, alongside hwtstamp-phyindex, to describe
the origin of the hardware timestamp.

Remove HWTSTAMP_SOURCE_UNSPEC enum value as it is not used at all.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
---
Change in v3:
- Move hwtstamp_source enum uapi/linux/ethtool.h to ethtool.yaml specs.
- Add few include of ethtool_netlink_generated.h header.
- Link to v2: https://lore.kernel.org/r/20250506-feature_ptp_source-v2-1-dec1c3181a7e@bootlin.com

Change in v2:
- Move hwtstamp_source enum from uapi/linux/net_tstamp.h to
  uapi/linux/ethtool.h
- Remove HWTSTAMP_SOURCE_UNSPEC enum value.
- Made a few change in net/ethtool/common.c to report the source only
  if get_ts_info does not return an error and if there is a hardware
  timestamp with a phc index.
- Link to v1: https://lore.kernel.org/r/20250425-feature_ptp_source-v1-1-c2dfe7b2b8b4@bootlin.com