devlink: Add devlink-conf uAPI for NV memory management
Introduce a new devlink-conf uAPI to manage device configuration stored
in non-volatile memory. This provides a standardized interface for
devices that need to persist configuration changes across reboots.

Add two new devlink operations:
- conf_save: Save current configuration to non-volatile memory
- conf_reset: Reset configuration in non-volatile memory to factory
	      defaults

The uAPI is designed to be generic and can be used by any device driver
that manages persistent configuration storage.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
---

After two NAKs on sysfs and devlink-param design choices due to uAPI not
being appropriate, I moved to this new design approach.
Also, the choice of saving the configuration to NVM on every config change
seemed inappropriate as the save and the reset processes would be in two
different Kernel framework.
The new design intend to extend devlink uAPI with a devlink-conf interface.
This uAPI is designed to be generic. While currently only one driver
implements it, controllers that manage non-volatile memory to save
configuration are not uncommon.

Change in v3:
- New patch.